+++ test/pushgatewayTest.js
@@ -3,21 +3,33 @@
 const nock = require('nock');
 const { gzipSync } = require('zlib');
 
-describe('pushgateway', () => {
+const Registry = require('../index').Registry;
+
+describe.each([
+	['Prometheus', Registry.PROMETHEUS_CONTENT_TYPE],
+	['OpenMetrics', Registry.OPENMETRICS_CONTENT_TYPE],
+])('pushgateway with %s registry', (tag, regType) => {
 	const Pushgateway = require('../index').Pushgateway;
 	const register = require('../index').register;
-	const Registry = require('../index').Registry;
 	let instance;
 	let registry = undefined;
 
+	beforeEach(() => {
+		register.setContentType(regType);
+	});
+
 	const tests = function () {
+		let body;
+		if (regType === Registry.OPENMETRICS_CONTENT_TYPE) {
+			body = '# HELP test test\n# TYPE test counter\ntest_total 100\n# EOF\n';
+		} else {
+			body = '# HELP test test\n# TYPE test counter\ntest 100\n';
+		}
+
 		describe('pushAdd', () => {
 			it('should push metrics', () => {
 				const mockHttp = nock('http://192.168.99.100:9091')
-					.post(
-						'/metrics/job/testJob',
-						'# HELP test test\n# TYPE test counter\ntest 100\n',
-					)
+					.post('/metrics/job/testJob', body)
 					.reply(200);
 
 				return instance.pushAdd({ jobName: 'testJob' }).then(() => {
@@ -27,10 +39,7 @@ describe('pushgateway', () => {
 
 			it('should use groupings', () => {
 				const mockHttp = nock('http://192.168.99.100:9091')
-					.post(
-						'/metrics/job/testJob/key/value',
-						'# HELP test test\n# TYPE test counter\ntest 100\n',
-					)
+					.post('/metrics/job/testJob/key/value', body)
 					.reply(200);
 
 				return instance
@@ -45,10 +54,7 @@ describe('pushgateway', () => {
 
 			it('should escape groupings', () => {
 				const mockHttp = nock('http://192.168.99.100:9091')
-					.post(
-						'/metrics/job/testJob/key/va%26lue',
-						'# HELP test test\n# TYPE test counter\ntest 100\n',
-					)
+					.post('/metrics/job/testJob/key/va%26lue', body)
 					.reply(200);
 
 				return instance
@@ -65,10 +71,7 @@ describe('pushgateway', () => {
 		describe('push', () => {
 			it('should push with PUT', () => {
 				const mockHttp = nock('http://192.168.99.100:9091')
-					.put(
-						'/metrics/job/testJob',
-						'# HELP test test\n# TYPE test counter\ntest 100\n',
-					)
+					.put('/metrics/job/testJob', body)
 					.reply(200);
 
 				return instance.push({ jobName: 'testJob' }).then(() => {
@@ -78,10 +81,7 @@ describe('pushgateway', () => {
 
 			it('should uri encode url', () => {
 				const mockHttp = nock('http://192.168.99.100:9091')
-					.put(
-						'/metrics/job/test%26Job',
-						'# HELP test test\n# TYPE test counter\ntest 100\n',
-					)
+					.put('/metrics/job/test%26Job', body)
 					.reply(200);
 
 				return instance.push({ jobName: 'test&Job' }).then(() => {
@@ -117,10 +117,7 @@ describe('pushgateway', () => {
 
 			it('pushAdd should send POST request with basic auth data', () => {
 				const mockHttp = nock(`http://${auth}@192.168.99.100:9091`)
-					.post(
-						'/metrics/job/testJob',
-						'# HELP test test\n# TYPE test counter\ntest 100\n',
-					)
+					.post('/metrics/job/testJob', body)
 					.reply(200);
 
 				return instance.pushAdd({ jobName: 'testJob' }).then(() => {
@@ -130,10 +127,7 @@ describe('pushgateway', () => {
 
 			it('push should send PUT request with basic auth data', () => {
 				const mockHttp = nock(`http://${auth}@192.168.99.100:9091`)
-					.put(
-						'/metrics/job/testJob',
-						'# HELP test test\n# TYPE test counter\ntest 100\n',
-					)
+					.put('/metrics/job/testJob', body)
 					.reply(200);
 
 				return instance.push({ jobName: 'testJob' }).then(() => {
@@ -158,10 +152,7 @@ describe('pushgateway', () => {
 					'unit-test': '1',
 				},
 			})
-				.put(
-					'/metrics/job/testJob',
-					'# HELP test test\n# TYPE test counter\ntest 100\n',
-				)
+				.put('/metrics/job/testJob', body)
 				.reply(200);
 
 			instance = new Pushgateway(
@@ -185,10 +176,7 @@ describe('pushgateway', () => {
 					'Content-Encoding': 'gzip',
 				},
 			})
-				.post(
-					'/metrics/job/testJob',
-					gzipSync('# HELP test test\n# TYPE test counter\ntest 100\n'),
-				)
+				.post('/metrics/job/testJob', gzipSync(body))
 				.reply(200);
 
 			instance = new Pushgateway(
@@ -229,7 +217,7 @@ describe('pushgateway', () => {
 		});
 
 		beforeEach(() => {
-			registry = new Registry();
+			registry = new Registry(regType);
 			instance = new Pushgateway('http://192.168.99.100:9091', null, registry);
 			const promeClient = require('../index');
 			const cnt = new promeClient.Counter({